+++ packages/sdk-router/src/swap/engines/index.ts
@@ -1,6 +1,7 @@
 export * from './defaultPoolsEngine'
 export * from './kyberSwapEngine'
 export * from './liFiEngine'
+export * from './liquidSwapEngine'
 export * from './noOpEngine'
 export * from './paraSwapEngine'
 export * from './response'